refactor: add jwt provider impl
