+++ src/main/java/br/com/thiagoRDS/api_authors/modules/authors/usecases/AuthenticateAuthor.java
@@ -8,7 +8,7 @@ import br.com.thiagoRDS.api_authors.modules.authors.dtos.AuthenticateAuthorRespo
 import br.com.thiagoRDS.api_authors.modules.authors.entities.Author;
 import br.com.thiagoRDS.api_authors.modules.authors.exceptions.InvalidCredentialsException;
 import br.com.thiagoRDS.api_authors.modules.authors.repositories.AuthorsRepository;
-import br.com.thiagoRDS.api_authors.providers.JwtProvider;
+import br.com.thiagoRDS.api_authors.providers.JwtProvider.JwtProvider;
 import br.com.thiagoRDS.api_authors.providers.dtos.SignResponseDTO;
 
 @Service
