+++ src/test/java/br/com/thiagoRDS/api_authors/modules/authors/controllers/UpdatePasswordControllerTest.java
@@ -25,7 +25,7 @@ import br.com.thiagoRDS.api_authors.modules.authors.entities.Author;
 import br.com.thiagoRDS.api_authors.modules.authors.repositories.AuthorsRepository;
 import br.com.thiagoRDS.api_authors.modules.utils.Convert;
 import br.com.thiagoRDS.api_authors.modules.utils.MakeAuthor;
-import br.com.thiagoRDS.api_authors.providers.JwtProvider;
+import br.com.thiagoRDS.api_authors.providers.JwtProvider.JwtProvider;
 import br.com.thiagoRDS.api_authors.providers.dtos.SignResponseDTO;
 
 @Transactional
