+++ src/test/java/br/com/thiagoRDS/api_authors/modules/posts/controllers/UpdatePostControllerTest.java
@@ -30,7 +30,7 @@ import br.com.thiagoRDS.api_authors.modules.posts.repositories.PostsRepository;
 import br.com.thiagoRDS.api_authors.modules.utils.Convert;
 import br.com.thiagoRDS.api_authors.modules.utils.MakeAuthor;
 import br.com.thiagoRDS.api_authors.modules.utils.MakePost;
-import br.com.thiagoRDS.api_authors.providers.JwtProvider;
+import br.com.thiagoRDS.api_authors.providers.JwtProvider.JwtProvider;
 import br.com.thiagoRDS.api_authors.providers.dtos.SignResponseDTO;
 
 @Transactional
